adding streaming listener
